feat(ingest): refactor to remove copy paste
