+++ metadata-ingestion/src/datahub/ingestion/transformer/clear_dataset_ownership.py
@@ -1,9 +1,7 @@
-from typing import Iterable
-
 import datahub.emitter.mce_builder as builder
 from datahub.configuration.common import ConfigModel
-from datahub.ingestion.api.common import PipelineContext, RecordEnvelope
-from datahub.ingestion.api.transform import Transformer
+from datahub.ingestion.api.common import PipelineContext
+from datahub.ingestion.transformer.dataset_transformer import DatasetTransformer
 from datahub.metadata.schema_classes import (
     DatasetSnapshotClass,
     MetadataChangeEventClass,
@@ -15,7 +13,7 @@ class ClearDatasetOwnershipConfig(ConfigModel):
     pass
 
 
-class SimpleClearDatasetOwnership(Transformer):
+class SimpleClearDatasetOwnership(DatasetTransformer):
     """Transformer that adds a specified set of owners to each dataset."""
 
     def __init__(self, config: ClearDatasetOwnershipConfig, ctx: PipelineContext):
@@ -27,14 +25,6 @@ class SimpleClearDatasetOwnership(Transformer):
     ) -> "SimpleClearDatasetOwnership":
         config = ClearDatasetOwnershipConfig.parse_obj(config_dict)
         return cls(config, ctx)
-
-    def transform(
-        self, record_envelopes: Iterable[RecordEnvelope]
-    ) -> Iterable[RecordEnvelope]:
-        for envelope in record_envelopes:
-            if isinstance(envelope.record, MetadataChangeEventClass):
-                envelope.record = self.transform_one(envelope.record)
-            yield envelope
 
     def transform_one(self, mce: MetadataChangeEventClass) -> MetadataChangeEventClass:
         if not isinstance(mce.proposedSnapshot, DatasetSnapshotClass):
